refactor: Separate component
